backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
firebase-admin==6.5.0
psycopg2-binary==2.9.9
pg8000==1.31.2
gunicorn==21.2.0
pytest==8.3.3
